Moved prediction service methods of collector to separate file
